Set up the watchOS2 sample.
So having the watchOS2 sample present prevents the OS1 sample from running on the simulator. I think the issue is that the simulator is running watchOS 2 and it gets confused. But honestly, I don't care that much. If you remove the watchOS2 targets from the app the OS1 sample runs fine, and if you add the watchOS2 targets then the watchOS2 targets run.

I suspect dropping back to a earlier simulator that ran watchOS1 would work fine but I'm not that interested right now.